requirements.txt
Django==1.8.14
psycopg2==2.6.2
django-extensions==1.7.2
django_compressor==2.1
django-allauth==0.27.0
django-modeltranslation==0.11